# specs/002-api-contract/openapi-v0.yaml
openapi: 3.1.0
info:
  title: ERP-ALL API
  version: 0.1.0-draft
  description: >
    契约草案 v0（EA-002）。五个核心上下文：identity / channel / catalog / listing / order + 采购门户。
    全局约定见同目录 README.md；数据语义以 specs/001-domain-model/ 为准。
    x-permission = 所需权限点（permission.code），网关中间件按此校验。
servers:
  - url: /api/v1
    description: 内部 API（JWT audience=erp）
  - url: /portal/v1
    description: 采购门户 API（JWT audience=portal，仅 Portal tag 下端点）
...
security:
  - bearerAuth: []
paths:
...
  /pricing-strategies:
    get:
      tags: [Listing]
      x-permission: pricing.read
      responses: {'200': {description: OK, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/PricingStrategy'}}}}}}
    post:
      tags: [Listing]
      summary: 建策略（(team×store×offer_mode) 活跃唯一，D-Q23）
      x-permission: pricing.write
      requestBody: {$ref: '#/components/requestBodies/PricingStrategyWrite'}
      responses: {'201': {description: Created}, '409': {$ref: '#/components/responses/Error'}}
  /pricing-strategies/{strategyId}:
    patch:
      tags: [Listing]
      x-permission: pricing.write
      parameters: [{name: strategyId, in: path, required: true, schema: {type: integer}}]
      requestBody: {$ref: '#/components/requestBodies/PricingStrategyWrite'}
      responses: {'200': {description: OK}}
  # R2-06 增量2 契约补充冻结：重定价预览（只读试算不落库）
  /pricing-strategies/preview:
    post:
      tags: [Listing]
      summary: 重定价预览（clamp 版展示区间外情况 + compute_price 严格判定并列返回；manual 策略 reason='manual'）
      x-permission: pricing.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [store_id, offer_mode]
              properties:
                store_id: {type: integer}
                offer_mode: {type: string, enum: [build, match]}
                product_ids: {type: array, items: {type: integer}, maxItems: 200, description: 与 listing_ids 二选一}
                listing_ids: {type: array, items: {type: integer}, maxItems: 200, description: 与 product_ids 二选一（join listing 取 product + current_price 作 old_price）}
      responses:
        '200':
          description: 试算结果（不落库）
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategy:
                    type: object
                    properties:
                      id: {type: integer}
                      name: {type: string}
                      algo_code: {type: string}
                      version: {type: integer}
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id: {type: integer}
                        listing_id: {type: [integer, 'null']}
                        old_price: {type: [number, 'null']}
                        new_price: {type: [number, 'null'], description: clamp 版参考价（区间外 detail.out_of_band=true）}
                        ok: {type: boolean, description: compute_price 严格判定（区间外/低于底线为 false）}
                        reason: {type: [string, 'null'], examples: [out_of_band, below_min_price, no_source_price, manual, not_found]}
                        detail: {type: object}
        '422': {description: 无活跃策略（PRICING_STRATEGY_NOT_FOUND）/入参二选一违规 等业务拒绝, $ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: 内部 JWT（audience=erp）}
    portalAuth: {type: http, scheme: bearer, bearerFormat: JWT, description: 门户 JWT（audience=portal）}
...
  responses:
    Error:
      description: 统一错误信封
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  requestBodies:
...
    PricingStrategyWrite:
      content:
        application/json:
          schema:
            type: object
            properties:
              store_id: {type: [integer, 'null']}
              offer_mode: {type: string, enum: [build, match]}
              name: {type: string}
              algo_code: {type: string}
              params: {type: object, description: 必含 min_price 硬底线}
              status: {type: string, enum: [active, disabled]}
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            detail: {type: object}
            request_id: {type: string}
...
    PricingStrategy:
      type: object
      properties:
        id: {type: integer}
        store_id: {type: [integer, 'null']}
        offer_mode: {type: string}
        name: {type: string}
        algo_code: {type: string}
        params: {type: object}
        status: {type: string}
        version: {type: integer}
...